Extract BitwardenPasswordProtectedFileFormat
