+++ libs/importer/src/importers/bitwarden/bitwarden-json-types.ts
@@ -1,0 +1,11 @@
+export interface BitwardenPasswordProtectedFileFormat {
+  encrypted: boolean;
+  passwordProtected: boolean;
+  salt: string;
+  kdfIterations: number;
+  kdfMemory?: number;
+  kdfParallelism?: number;
+  kdfType: number;
+  encKeyValidation_DO_NOT_EDIT: string;
+  data: string;
+}
